requirements.txt
anthropic>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
playwright>=1.40.0
nest-asyncio>=1.5.8
pandas>=2.0.0
Pillow>=10.0.0
pydantic>=2.0.0
tabulate>=0.9.0
ipython>=8.0.0
